BpkRating to TS and remove className usage on BpkText
