+++ package.json
@@ -137,6 +137,7 @@
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.12",
     "@types/jest-axe": "^3.5.3",
+    "@types/lodash.clamp": "^4.0.9",
     "@types/react": "^17.0.37",
     "@types/react-dom": "^17.0.11",
     "@types/react-window": "^1.8.8",
