+++ packages/bpk-component-rating/index.ts
