+++ packages/bpk-component-rating/src/common-types.ts
@@ -15,24 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/* @flow strict */
 
-import { render } from '@testing-library/react';
-import { axe } from 'jest-axe';
+export const RATING_SIZES = { base: 'base', large: 'large' } as const;
 
-import BpkRating from '../index';
-
-describe('BpkRating accessibility tests', () => {
-  it('should not have programmatically-detectable accessibility issues', async () => {
-    const { container } = render(
-      <BpkRating
-        ariaLabel="3.6 Average might recommend"
-        title="Average"
-        subtitle="Might recommend"
-        value={3.6}
-      />,
-    );
-    const results = await axe(container);
-    expect(results).toHaveNoViolations();
-  });
-});
+export const RATING_SCALES = {
+  zeroToFive: 'zeroToFive',
+  zeroToTen: 'zeroToTen',
+} as const;
